+++ sample-project/lib/src/routes/controllers/controllers.dart
@@ -1,0 +1,8 @@
+library sample_project.src.routes.controllers;
+
+import 'dart:async';
+import 'package:angel_framework/angel_framework.dart';
+
+Future configureServer(Angel app) async {
+  /// Controllers will not function unless wired to the application!
+}